Add check for correct user to delete a book

+++ deletebook.php
@@ -12,6 +12,17 @@
 
         // Escape special characters in a string
         $id = mysqli_real_escape_string($conn, $id);
+
+        // Check if the book belongs to the user
+        $sql = "SELECT * FROM books WHERE id = $id";
+        $result = mysqli_query($conn, $sql);
+        $row = mysqli_fetch_assoc($result);
+        echo $row['user_id']; 
+        echo $_SESSION['user_id']; 
+        if ($_SESSION['user_id'] != $row['user_id']) { 
+            header('location: viewallbooks.php');
+            die; 
+        }
         
         // Create a SQL query
         $sql = "DELETE FROM books WHERE id = $id";
